fix: lp confirm and status reusable (#6055)
* composable

* dummy logic for status

* Update ReusableLpConfirm.tsx

* keys

* turn into a switch

* Update ReusableLpConfirm.tsx

* clean up

* move the tx card logic into the reusable component

* Update ReusableLpStatus.tsx

* Update ReusableLpStatus.tsx
